Added TS support and config
