--- components/head.js
@@ -1,50 +1,0 @@
-import React from "react";
-import NextHead from "next/head";
-
-const Head = ({ title, description }) => (
-	<NextHead>
-		<meta charset="UTF-8" />
-		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-		<meta http-equiv="X-UA-Compatible" content="ie=edge" />
-		<title>Female Freelance Developers</title>
-		<link rel="stylesheet" href="style.css" />
-		<script src="script.js" />
-		<link rel="icon" type="image/png" href="icon.png" />
-		<meta charset="UTF-8" />
-		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-		<meta http-equiv="X-UA-Compatible" content="ie=edge" />
-		<meta
-			name="description"
-			content="Get a more diverse selection of cool people to hire for your projects. Presented by Cafe Robot and curated by Valerie Fuchs."
-		/>
-
-		<meta property="og:url" content="https://femalefreelancedevelopers.com/" />
-		<meta property="og:type" content="website" />
-		<meta property="og:title" content="Female Freelance Developers" />
-		<meta
-			property="og:description"
-			content="Get a more diverse selection of cool people to hire for your projects. Presented by Cafe Robot and curated by Valerie Fuchs."
-		/>
-		<meta
-			property="og:image"
-			content="https://femalefreelancedevelopers.com/og_image.jpg"
-		/>
-
-		<meta
-			property="twitter:url"
-			content="https://femalefreelancedevelopers.com/"
-		/>
-		<meta name="twitter:card" content="summary_large_image" />
-		<meta name="twitter:title" content="Female Freelance Developers" />
-		<meta
-			name="twitter:description"
-			content="Get a more diverse selection of cool people to hire for your projects. Presented by Cafe Robot and curated by Valerie Fuchs."
-		/>
-		<meta
-			name="twitter:image"
-			content="https://femalefreelancedevelopers.com/og_image.jpg"
-		/>
-	</NextHead>
-);
-
-export default Head;
