+++ package.json
@@ -14,5 +14,10 @@
   },
   "engines": {
     "node": "22.x"
+  },
+  "devDependencies": {
+    "@types/node": "22.13.13",
+    "@types/react": "19.0.12",
+    "typescript": "5.8.2"
   }
 }
